+++ core/src/com/fourbit/pc_invader/utils/Globals.java
@@ -1,0 +1,7 @@
+package com.fourbit.pc_invader.utils;
+
+public class Globals {
+    public static final int GAME_WIDTH = 1920;  // measured in pixels
+    public static final int GAME_HEIGHT = 1080;  // Measured in pixels
+    public static final float PPM = 20;  // Pixel-per-meter ration
+}
